审核凭证没有权限时，没有提示； 删除银行现金时，如果没有权限，不显示403数字，只显示中文； 权限控制时，检测被编辑用户是否和当前登录用户在同一个组； 账号权限，账套所有者的权限不能修改； 账套没有权限时，账套首页也不能查看;

+++ protected/controllers/PermissionController.php
@@ -41,7 +41,37 @@ class PermissionController extends Controller
      */
     public function actionAccess($id)
     {
+
+        //不能修改所有者权限
+        $condom = Condom::model()->findByAttributes(['dbname'=>substr(SYSDB,8)]);
+        if ($condom->owner == $id) {
+            Yii::app()->user->setFlash('error', '不能修改账套所有者的权限！');
+            $this->redirect($this->createUrl('permission/index'), array());
+            return;
+        }
+
+        //不能修改自己
+        if (Yii::app()->user->id == $id) {
+            Yii::app()->user->setFlash('error', '不能修改自己的权限！');
+            $this->redirect($this->createUrl('permission/index'), array());
+            return;
+        }
+
+        //是否在同一个组
+        $user = User2::model()->findByPk(Yii::app()->user->id);
+        $g_users = User2::model()->findAllByAttributes(['group' => $user->group]);
+        $g_users_arr = [];
+        foreach ($g_users as $g_user) {
+            $g_users_arr[] = $g_user->id;
+        }
+        if(!in_array($id, $g_users_arr)) {
+            Yii::app()->user->setFlash('error', "该用户ID不在同一组中，请检查!");
+            $this->redirect($this->createUrl('permission/index'), array());
+            return;
+        }
+
         $model = AuthRelation::model()->findAllByAttributes(['user_id' => $id], ['order' => 'permission']);
+
         if (isset($_POST['AuthPermission'])) {
             AuthRelation::model()->deleteAllByAttributes(['user_id' => $id]);
             foreach ($_POST['AuthPermission'] as $item => $value) {
